fix role destory after edit user

+++ app/Http/Controllers/UserController.php
@@ -139,9 +139,11 @@ class UserController extends Controller
     
         $user = User::find($id);
         $user->update($input);
-        DB::table('model_has_roles')->where('model_id',$id)->delete();
-    
-        $user->assignRole($request->input('roles'));
+
+        if(Auth::user()->hasRole('Admin')){
+            DB::table('model_has_roles')->where('model_id',$id)->delete();
+            $user->assignRole($request->input('roles'));
+        }
     
         return redirect()->route('users.index')
                         ->with('success','User updated successfully');
